requirements.txt
aiogram==2.25.1
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
gino==1.0.1
SQLAlchemy==1.3.24
loguru==0.7.0
asyncpg==0.28.0
lxml
beautifulsoup4
html5lib
fake_useragent
pip==23.1.2
emoji==2.6.0
attrs==23.1.0
Pillow==10.0.0
Mako==1.2.4
Babel==2.9.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
alembic==1.11.3
setuptools==65.5.0
APScheduler==3.10.4